move BuiltinGoal to pants.goal.builtin_goal and registration to pants.goal.builtins
